+++ repro/models/gupta2020/__init__.py
@@ -2,8 +2,8 @@ import os
 
 VERSION = "1.0"
 MODEL_NAME = os.path.basename(os.path.dirname(__file__))
-DOCKERHUB_REPRO = f"danieldeutsch/{MODEL_NAME}"
-DEFAULT_IMAGE = f"{DOCKERHUB_REPRO}:{VERSION}"
+DOCKERHUB_REPO = f"danieldeutsch/{MODEL_NAME}"
+DEFAULT_IMAGE = f"{DOCKERHUB_REPO}:{VERSION}"
 AUTOMATICALLY_PUBLISH = True
 
 from repro.models.gupta2020.model import NeuralModuleNetwork