+++ Changelog.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Added [Benepar](models/kitaev2019/Readme.md)
+- Added [Lite3Pyramid](models/zhang2021/Readme.md)
+
+### Changed
+- Fixed silly variable name typo: `DOCKERHUB_REPRO` to `DOCKERHUB_REPO`
 
 ## [v0.1.0](https://github.com/danieldeutsch/repro/releases/tag/v0.1.0) - 2021-08-10
 ### Added